+++ packages/noise-cancellation-react-native/ios/Headers/NoiseCancellationManagerObjc.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface NoiseCancellationManagerObjc : NSObject
+
++ (instancetype)sharedInstance;
+- (void)registerProcessor;
+
+@end
+
+NS_ASSUME_NONNULL_END